Added functionality
Users can now choose when (or if) to see the graph, and can choose to run the program with the same settings (with the output going to a different file or Excel sheet). The latter feature will save time if multiple runs with the same settings are needed in one sitting.